+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Adds a SuperTokens + Supabase example app
 -   Adds an example app with svelte
+-   Adds phone number and password demo app
+-   Adds an example app with Angular + React
 
 ### CI changes